refactor(dicom-web): remove settings is-open, rename stores
Move formatBytes to utils

Formatting
