+++ src/utils/index.ts
@@ -138,3 +138,16 @@ export function wrapInArray<T>(maybeArray: T | T[]): T[] {
 export function getURLBasename(url: string) {
   return new URL(url).pathname.split('/').at(-1) ?? '';
 }
+
+// from https://stackoverflow.com/a/18650828
+export function formatBytes(bytes: number, decimals = 2) {
+  if (!+bytes) return '0 Bytes';
+
+  const k = 1024;
+  const dm = decimals < 0 ? 0 : decimals;
+  const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
+
+  const i = Math.floor(Math.log(bytes) / Math.log(k));
+
+  return `${parseFloat((bytes / k ** i).toFixed(dm))} ${sizes[i]}`;
+}
